Removing the elm.json file already present
I'm removing the elm.json file already present in the repo.
Instead, I instruct the user to run `elm init` when starting the project.
This will create an elm.json file that is compatible with the version of Elm present on the user's machine.
I'm also improving other parts of the README.rst.